+++ aasrp/views.py
@@ -140,16 +140,20 @@ def ajax_dashboard_srp_links_data(
                 aar_link=srp_link.aar_link, link_text=_("Link")
             )
 
-        srp_code_html = (
-            srp_link.srp_code + "<i "
-            'class="aa-srp-fa-icon aa-srp-fa-icon-right copy-text-fa-icon far fa-copy" '
-            'data-clipboard-text="{srp_link}" title="{title}"></i>'.format(
+        srp_code_html = srp_link.srp_code
+        if srp_link.srp_status == AaSrpStatus.ACTIVE:
+            srp_code_html += (
+                '<i class="{css_classes}" '
+                'data-clipboard-text="{srp_link}" title="{title}"></i>'
+            ).format(
+                css_classes=(
+                    "aa-srp-fa-icon aa-srp-fa-icon-right copy-text-fa-icon far fa-copy"
+                ),
                 srp_link=reverse_absolute(
                     "aasrp:request_srp", args=[srp_link.srp_code]
                 ),
                 title=_("Copy SRP link to clipboard"),
             )
-        )
 
         actions = get_dashboard_action_icons(request=request, srp_link=srp_link)
 
@@ -161,7 +165,6 @@ def ajax_dashboard_srp_links_data(
                 "fleet_commander": srp_link.fleet_commander.character_name,
                 "fleet_doctrine": srp_link.fleet_doctrine,
                 "aar_link": aar_link,
-                # "srp_code": srp_link.srp_code,
                 "srp_code": {"display": srp_code_html, "sort": srp_link.srp_code},
                 "srp_costs": srp_link.total_cost,
                 "srp_status": srp_link.srp_status,